Introspection expiration fixed, code cleanup
